fix(rpc): address review — search termination guard, gas-vector note, isolated nonce tests
