+++ madara/crates/client/rpc/src/test_utils.rs
@@ -110,6 +110,60 @@ pub async fn rpc_test_setup_with_execution() -> (Arc<MadaraBackend>, Starknet, m
     (backend, rpc, keys)
 }
 
+/// A fee-token transfer of 1 wei from `account` (a v3 invoke), signed iff `valid_signature`.
+/// For use with [`rpc_test_setup_with_execution`].
+#[cfg(test)]
+pub fn devnet_transfer_tx(
+    backend: &MadaraBackend,
+    account: &mc_devnet::DevnetPredeployedContract,
+    nonce: Felt,
+    valid_signature: bool,
+) -> mp_rpc::v0_9_0::BroadcastedTxn {
+    use mc_devnet::{Call, Multicall, Selector};
+    use mp_convert::ToFelt;
+    use mp_rpc::v0_9_0::{
+        BroadcastedInvokeTxn, BroadcastedTxn, DaMode, InvokeTxnV3, ResourceBounds, ResourceBoundsMapping,
+    };
+    use mp_transactions::IntoStarknetApiExt;
+
+    let mut tx = InvokeTxnV3 {
+        sender_address: account.address,
+        calldata: Multicall::default()
+            .with(Call {
+                to: backend.chain_config().native_fee_token_address.to_felt(),
+                selector: Selector::from("transfer"),
+                calldata: vec![account.address, Felt::ONE, Felt::ZERO],
+            })
+            .flatten()
+            .collect::<Vec<_>>()
+            .into(),
+        signature: vec![Felt::ONE, Felt::TWO].into(),
+        nonce,
+        resource_bounds: ResourceBoundsMapping {
+            l1_gas: ResourceBounds { max_amount: 60000, max_price_per_unit: 10000 },
+            l2_gas: ResourceBounds { max_amount: 6000000000, max_price_per_unit: 100000 },
+            l1_data_gas: ResourceBounds { max_amount: 60000, max_price_per_unit: 10000 },
+        },
+        tip: 0,
+        paymaster_data: vec![],
+        account_deployment_data: vec![],
+        nonce_data_availability_mode: DaMode::L1,
+        fee_data_availability_mode: DaMode::L1,
+    };
+    if valid_signature {
+        let api_tx = BroadcastedTxn::Invoke(BroadcastedInvokeTxn::V3(tx.clone()))
+            .into_validated_tx(
+                backend.chain_config().chain_id.to_felt(),
+                backend.chain_config().latest_protocol_version,
+                TxTimestamp::now(),
+            )
+            .unwrap();
+        let signature = account.secret.sign(&api_tx.hash).unwrap();
+        tx.signature = vec![signature.r, signature.s].into();
+    }
+    BroadcastedTxn::Invoke(BroadcastedInvokeTxn::V3(tx))
+}
+
 #[fixture]
 pub fn rpc_test_setup() -> (Arc<MadaraBackend>, Starknet) {
     let chain_config = Arc::new(ChainConfig::madara_test());
