+++ madara/crates/client/rpc/src/versions/user/v0_7_1/methods/trace/simulate_transactions.rs
@@ -41,7 +41,10 @@ pub async fn simulate_transactions(
 
     // spawn_blocking: avoid starving the tokio workers during execution.
     let (execution_results, exec_context) = mp_utils::spawn_blocking(move || {
-        Ok::<_, mc_exec::Error>((exec_context.execute_transactions_for_estimation([], user_transactions)?, exec_context))
+        Ok::<_, mc_exec::Error>((
+            exec_context.execute_transactions_for_estimation([], user_transactions)?,
+            exec_context,
+        ))
     })
     .await
     .map_err(StarknetRpcApiError::from_exec_error_v0_7)?;
